+++ src/commands/russian_roulette.py
@@ -19,6 +19,7 @@ TABLES_ALIASES = {
     'луз': 'lose',
     'ноль': 'zero',
     'минус': 'minus'}
+ADMIN_LIST = get_data_from_database(0, 'admin_list', 'admins_id')
 DELAY_TIME = 3
 
 
@@ -80,17 +81,18 @@ async def start_roulette(msg, args):
                 await msg.delete()
             return
         if args[0] == 'удалить':
-            if args[1] in TABLES_ALIASES:
-                TABLE_TO_MODIFY = TABLES_ALIASES[args[1]]
-                for i in range(2):
-                    args.pop(0)
-                WORD_TO_DELETE = " ".join(args)
-                await msg.channel.send(delete_roulette_word(
-                                       WORD_TO_DELETE, TABLE_TO_MODIFY
-                                       ),
-                                       delete_after=DELAY_TIME)
-                await asyncio.sleep(DELAY_TIME)
-                await msg.delete()
+            if msg.author.id in ADMIN_LIST:
+                if args[1] in TABLES_ALIASES:
+                    TABLE_TO_MODIFY = TABLES_ALIASES[args[1]]
+                    for i in range(2):
+                        args.pop(0)
+                    WORD_TO_DELETE = " ".join(args)
+                    await msg.channel.send(delete_roulette_word(
+                                        WORD_TO_DELETE, TABLE_TO_MODIFY
+                                        ),
+                                        delete_after=DELAY_TIME)
+                    await asyncio.sleep(DELAY_TIME)
+                    await msg.delete()
             return
 
         try:
